requirements.txt
streamlit
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client 
pandas